pyproject.toml
[tool.poetry]
name = "forum-engine"
version = "0.1.0"
description = ""
authors = ["ybelmach <yaroslav.belmach.wk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.35"
fastapi = "^0.115.0"
python-dotenv = "^1.0.1"
psycopg = "^3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
